hello

+++ installer.sh
@@ -1,2 +1,11 @@
 sudo echo  "alias serve='bash $PWD/index.sh'" >> "~/.bash_profile"
-sudo rm "$PWD/installer.sh"
+echo "Delete installer?
+Y|n
+"
+read r
+if [[ "$r" == "y" || "$r" == "yes"  || "$r" == "Y" ]]; then
+ sudo rm "$PWD/installer.sh"
+fi
+echo "Installation done
+Run 'server help'
+"
